pyproject.toml
[project]
name = "lmgate"
version = "0.1.0"
description = "Transparent pass-through proxy for LLM API providers with access control and usage statistics"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9,<4",
    "pyyaml>=6,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-aiohttp>=1.0,<2",
    "aioresponses>=0.7,<1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["lmgate*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
